add comment function

+++ src/main/java/com/community/controller/PublishController.java
@@ -33,7 +33,7 @@ public class PublishController {
     public String publish(@RequestParam(name="title", required = false) String title,
                           @RequestParam(name="description", required = false) String description,
                           @RequestParam(name="tag", required = false) String tag,
-                          @RequestParam(name = "id", required = false) Integer id,
+                          @RequestParam(name = "id", required = false) Long id,
                           HttpServletRequest request,
                           Model model) {
 
@@ -64,7 +64,7 @@ public class PublishController {
         question.setTitle(title);
         question.setTag(tag);
         question.setDescription(description);
-        question.setCreator(Integer.valueOf(user.getId()));
+        question.setCreator(user.getId());
         question.setGmtCreate(System.currentTimeMillis());
         question.setGmtModified(question.getGmtCreate());
         question.setId(id);
@@ -74,7 +74,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
 
         QuestionDTO question = questionService.getQuestionById(id);
